[openrr-planner] Move a test of CollisionDetector to the suitable file
